investigacion_api/investigacionApp/views.py
from django.shortcuts import render
from rest_framework import permissions, viewsets

from .models import Persona
from .serializers import PersonaSerializer


class PersonaViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows users to be viewed or edited.
    """
    queryset = Persona.objects.all()
    serializer_class = PersonaSerializer
    permission_classes = [permissions.IsAuthenticated]


def index(request):
    return render(request, "index.html")
