requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
